+++ tests/Unit/Migration/Templates/MigrationTemplateTest.php
@@ -11,7 +11,8 @@ class MigrationTemplateTest extends AppTestCase
     {
         $template = MigrationTemplate::create('create_table_users_1001', 'users');
 
-        $this->assertStringContainsString('class Create_table_users_1001 extends QtMigration', $template);
+        $this->assertStringContainsString('use Quantum\Migration\Migration;', $template);
+        $this->assertStringContainsString('class Create_table_users_1001 extends Migration', $template);
         $this->assertStringContainsString("\$tableFactory->create('users')", $template);
         $this->assertStringContainsString("\$tableFactory->drop('users')", $template);
     }
@@ -20,7 +21,8 @@ class MigrationTemplateTest extends AppTestCase
     {
         $template = MigrationTemplate::alter('alter_table_users_1002', 'users');
 
-        $this->assertStringContainsString('class Alter_table_users_1002 extends QtMigration', $template);
+        $this->assertStringContainsString('use Quantum\Migration\Migration;', $template);
+        $this->assertStringContainsString('class Alter_table_users_1002 extends Migration', $template);
         $this->assertStringContainsString("\$tableFactory->get('users')", $template);
     }
 
@@ -28,7 +30,8 @@ class MigrationTemplateTest extends AppTestCase
     {
         $template = MigrationTemplate::rename('rename_table_users_1003', 'users');
 
-        $this->assertStringContainsString('class Rename_table_users_1003 extends QtMigration', $template);
+        $this->assertStringContainsString('use Quantum\Migration\Migration;', $template);
+        $this->assertStringContainsString('class Rename_table_users_1003 extends Migration', $template);
         $this->assertStringContainsString("\$tableFactory->rename('users', \$newName)", $template);
         $this->assertStringContainsString("\$tableFactory->rename(\$newName, 'users')", $template);
     }
@@ -37,7 +40,8 @@ class MigrationTemplateTest extends AppTestCase
     {
         $template = MigrationTemplate::drop('drop_table_users_1004', 'users');
 
-        $this->assertStringContainsString('class Drop_table_users_1004 extends QtMigration', $template);
+        $this->assertStringContainsString('use Quantum\Migration\Migration;', $template);
+        $this->assertStringContainsString('class Drop_table_users_1004 extends Migration', $template);
         $this->assertStringContainsString("\$tableFactory->drop('users')", $template);
     }
 }